requirements.txt
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cymem==2.0.5
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
en-core-web-lg==2.3.1
en-core-web-sm==2.3.1
entrypoints==0.3
fasttext==0.9.2
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.7.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.19.4
packaging==20.7
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.10
pybind11==2.6.1
pycparser==2.20
Pygments==2.7.3
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
scikit-learn==0.24.1
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==4.0.1
smmap==3.0.5
spacy==2.3.4
srsly==1.0.5
terminado==0.9.1
testpath==0.4.4
thinc==7.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
